requirements.txt
python-dotenv~=1.0.0
typing~=3.7.4.3
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pytz~=2024.1
discord.py
asyncio~=3.4.3